I finished the Date and ValidDate classes, along with relevant exceptions and test cases. I don't know where to put the test cases, so someone else is going to need to move them.
The testing is not complete, in that it doesn't check all the Date
examples when trying ValidDate. I did this because it would work exactly
the same way, it would be really annoying to change all the years to be
in the appropriate range, and the test would only work properly after a
certain day. This is especially bad since the first loop leap year after
1999 is 2100.